+++ src/main/java/fr/univ_cote_azur/ai_game_programming/Thread/Task.java
@@ -9,9 +9,9 @@ public class Task implements Runnable {
     boolean isMax;
     int depth;
     int[] move;
-    private float local_evaluation;
+    private double local_evaluation;
 
-    public Task(int[][] board, int turn, float eval_Parent, boolean isMax, int depth, int[] move) {
+    public Task(int[][] board, int turn, double eval_Parent, boolean isMax, int depth, int[] move) {
         this.board = new int[3][16];
         arraysOperations.deepCopy(board, this.board);
         this.turn = turn;
@@ -30,11 +30,11 @@ public class Task implements Runnable {
     }
 
 
-    public synchronized float minMax(int[][] parent_board, int turn, int[] parent_move, float parent_eval, boolean isMax, int depth) {
+    public synchronized double minMax(int[][] parent_board, int turn, int[] parent_move, double parent_eval, boolean isMax, int depth) {
 
         int[][] local_board = new int[3][16];
         arraysOperations.deepCopy(parent_board, local_board);
-        float local_eval = parent_eval;
+        double local_eval = parent_eval;
         // Simulate the parent play here.
         Simulate_Player player = new Simulate_Player((turn + 1) % 2);
         player.setScore(0);
@@ -49,7 +49,7 @@ public class Task implements Runnable {
             if (!isMax) {
                 int count_criticHoles = arraysOperations.count_criticHoles(local_board, (turn + 1) % 2);
                 // 1/16 = 0.0625
-                if (count_criticHoles > 0) local_eval -= (float) (0.0620 * count_criticHoles);
+                if (count_criticHoles > 0) local_eval -= 0.0620 * count_criticHoles;
             }
             return local_eval;
         }
@@ -61,14 +61,14 @@ public class Task implements Runnable {
             return 100;
         }
 
-        float bestEval;
+        double bestEval;
         if (isMax) bestEval = -100;
         else bestEval = 100;
 
 
         for (int[] move : legitMoves) {
 
-            float score = minMax(local_board, (turn + 1) % 2, move, local_eval, !isMax, depth - 1);
+            double score = minMax(local_board, (turn + 1) % 2, move, local_eval, !isMax, depth - 1);
 
             if (move == legitMoves[0]) {
                 bestEval = score;
@@ -84,13 +84,13 @@ public class Task implements Runnable {
         return bestEval;
     }
 
-    protected synchronized float eval(boolean isMax, float local_parent_eval, float score) {
+    protected synchronized double eval(boolean isMax, double local_parent_eval, double score) {
         if (isMax && score > local_parent_eval) return score;
         if (!isMax && score < local_parent_eval) return score;
         return local_parent_eval;
     }
 
-    public float getEval() {
+    public double getEval() {
         return local_evaluation;
     }
 }
